refactor(backend): isolate invite redemption transaction
